requirements.txt
langgraph>=0.2.0
langsmith>=0.1.0
langgraph-checkpoint-sqlite>=2.0.0
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
slack-sdk>=3.33.0
pydantic>=2.9.0
pytest>=8.0.0
structlog>=24.0.0
